# tabeeby-healthcare-os/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: TABEEBY Healthcare OS API
  version: 1.0.0
  description: |
    The world's most advanced medical AI platform API.
    Ultra IQ 300B Cognitive Engine | 20M+ concurrent users
servers:
  - url: https://api.tabeeby.health/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local Development
paths:
...
  /patients/register:
    post:
      summary: Register patient with lifetime AI physician
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                date_of_birth: { type: string }
                gender: { type: string }
                language: { type: string }
      responses:
        200:
          description: Patient registered
...